+++ ICanSeeClearlyNow/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 0.0.3
+
+  * Changed mod behavior to only be active during the in game night and not all the time
+  * Modified support for zones that are always dark (dungeons, the swamps, etc.) to be in line with other options
+  * Added custom Start of Day and Start of Night options to allow for custom start/stop times
+  * Added toggle to enable/disable custom day and night timers
+  * Bumped up <LangVersion> to C# 10
+
 ### 0.0.2
 
   * Initial release of ICanSeeClearlyNow, Comfy Specific fork of Coldspirit's GammaMod.
@@ -10,4 +18,4 @@
 
 ### 0.0.1
 
-  * Initial release of GammaMod (ValheimNightGamma) by Coldspirit (https://github.com/ColdSpirit0/ValheimGammaMod).
+  * Initial release of GammaMod by Coldspirit (https://github.com/ColdSpirit0/ValheimGammaMod).